refactor: LargeReview 관련 파일 위치 common으로 변경
